+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 - 6.5.1
-  - Fix missing config file in antd moment date widgets (PR #1046)
+  - Fix missing config file in antd moment date widgets (PR #1046) (issue #1048)
 - 6.5.0
   - Breaking: `Utils.validateTree` returns array of errors (PR #1034)
   - Breaking: `removeIncompleteRulesOnLoad` by default is false